+++ processor/src/test/fixtures/input/com/example/component_dependency/ParameterizedFieldDependencyModel.java
@@ -1,0 +1,12 @@
+package com.example.component_dependency;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentDependency;
+
+@ArezComponent
+abstract class ParameterizedFieldDependencyModel<T>
+{
+  @ComponentDependency( validateTypeAtRuntime = true )
+  final T value = null;
+}
+